fixup! kernel/build: Add support for fragmented kernel configs
If the intention is to use fragmented configs, the maintainer should define the path to the fragmented config from the device side anyway. So, this is unnecessary and the logic is broken.
Consequently, as the logic is broken, if the maintainer does not use fragmented configs, leaving KERNEL_FRAGMENT_CONFIG empty, it will also fail causing a build error.

Fixes: 28fc89fa2b0e43878822abf51f3c978a14b4d434
Change-Id: I9380cc615cd23436c6ac13bf8443db7b03fc16b6
Signed-off-by: hpnightowl <iamhp2k@gmail.com>